+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 - Adding the `list--edit` icon to the icons preview
+- Added the `ToggleButton` component
+- Added the `HCLSoftwareBlue 07 (20P)` color
+- Added the `selectedOpacityHover` in theme.pallette.action
 
 ### Fixed